inc: remove minimum RAM warning
Either the RAM requirements have increased, or hardware has
increased.  Simply stating the arbitrary and obvious while
doing things of little value to `fix' the situation it is
probably better to just force an old-fashioned out of memory,
which is mostly the point where the user realises that it is
not going to work anyway.