Basic English .pot file, first formatting pass complete.
Warning: This file does not represent the last month of changes in the OPNsense codebase, next step is to follow up with an update to changed lines.  It is merely a first pass from-scratch dump from the source code.

Another Warning: many lines still contain errors, based on relatively inconsistent use of gettext calls in the source code.

There are many noted 'TODO' comments in the file, many of which in code which may just get ripped out soon- so keep track of current OPNsense core development efforts before you go too wild fixing something- it may be rewritten and your work would be wasted!

Signed-off-by: Isaac (.ike) Levy <ike@blackskyresearch.net>