dashboard: explicitly land on dashboard
The login redirects to index.php, but the ACL doesn't know about `/',
which is unfortunate because otherwise we can't land on the dashbard
anymore even though we have the rights.

While there, consolidate the ACL description for the dashbard.  It was
status, not system, but now it is lobby anyway...