firewall: always display all known gateways on edit
There are a few cases where the gateway code is eligible
to drop a gateway on grounds of being disabled, inactive,
unconfigured or where the actual interface assigned is
disabled.  This is undesire because users that configure
a gateway won't be able to use it.

There are no bad side effects from the former by selecting
a whatsoever dysfunctional gateway, because the problem still
has to be fixed by the user in the gateway setup itself.

Worst of all, removing a disabled but used gateway makes
it not show up on edit, effectively wiping its setting on
save.  This breaks POLA.

While there, also enable the null routes for IPv4 and IPv6.

PR: https://forum.opnsense.org/index.php?topic=2100