xmlrpc: prevent undefined errors on remote service restart
Further proof that all of the system should use the plugin system,
as the dependencies are going to be unmanageable in the long run.
As soon as there is a new base service with its own include file
its definition will be missing.

(cherry picked from commit f441eae5b45ce21a4ce1609b7464976dbbf8763f)