(model) refactor getConstraintValidators() in BaseField, add getConstraintByName() to retrieve a constraint by name/key and add functionality to explain the model field relations.
Using reference tags in the constraints section its possible to tell the model which fields interact to avoid the need for a full model validation after each change.
For example if fieldA has a constraint which relates to fieldB, the latter can add a reference to trigger validation when only fieldB changes.

<Constraints>
   <check001>
     <reference>fieldB.check000</reference>
   </check001>
</Constraints>