rc: revamp rc.loader
We only merge chunks for loader.conf leaveraging a priority scheme
in order to be able to override standard modifications made by e.g.
core or plugins (in case of user changes).

Start to work on its scripts, but 20-console is dynamic so it needs
more work in the actual PHP code.

Discussed with: @adschellevis