dhcp: register domain-name per subnet to prevent wrong default
According to manuals and other project's code this should work as
expected, but to make sure we just advertise the name without a
default.

While there, clean up the config file generation and a few printf
bits that really shouldn't be in there.

PR: https://github.com/opnsense/core/issues/1250