rc: don't (re)set shell privileges on password reset anymore
The admins group has this now, it makes more sense there as several
people ran into the issue of "my user is admin, but cannot login".

Besides, the privilege sets /bin/csh as the user shell, which we don't
use for root as we have /usr/local/etc/rc.initial there.