mvc, model. Validation issue when storing a value for the first time.
If a field isn't required and the user sets it with a faulty value, it would pass because it didn't know it wasn't set before user input.
There where two issues, first one, not knowing it was in init state (solved with internalFieldLoaded), next all field types should have used actionPostLoadingEvent to customize loading behaviour.