wizard: do not force change of root password
We are forcing a change through the installer, which is where
the majority starts from.  We don't want to enforce a change
twice.

We can't really see if we installed or went through a factory
reset, so this is (was) the best of both worlds in conjunction
with the new installer feature.

While here, annotate that it is possible to leave it "empty".

This reverts commit 6882b36984e475fead11c2e9b6d4c0a1a1c79e43.

(cherry picked from commit 9cf6655b46bba17e62bdbf8408ffc739de30f265)