firmware: automatically support "other" flavours
We always have an internal ABI prefix now, if the flavour is a short
keyword, we should always use the ABI on it so "experimental" becomes
"17.7/experimental".  If a path is set "17.7/MINT/17.7/LibreSSL" we
detect the slash and assume the explicit ABI was selected.

This may eventually support things such as "nightly" or others without
fiddling through this code every time something else is added.