HTML Compliance - MVC Form Input - Attribute "for" on Element <tr,div,a,small,span>
/usr/core/src/mvc/app/views/layout_partials/form_input_tr.volt
Error: Attribute "for' not allowed on element <tr> at this point.
Error: Attribute "for" not allowed on element <div> at this point.
Error: Attribute "for" not allowed on element <a> at this point.
Error: Attribute "for" not allowed on element <small> at this point.
Error: Attribute "for" not allowed on element <span> at this point.

Error: Duplicate ID "clear-options".