unbound: optimize outgoing-range differently; closes #2172
"- 50" would will eventually underflow.  According to the official
docs in the libevent case we set double num-queries-per-thread, so
that's what we'll do and kick the other code.

While here, seems like the GUI options for the default value are
missing which causes the num-queries-per-thread value to decrease
to 512, reducing the outbound-range with it arbitrarily.

See: https://www.unbound.net/documentation/howto_optimise.html