firmware: only check once, adhere to locks from console #2283
In the GUI it's marginally useful to say "your kernel / base needs
an update" but the update status says "it was locked".  In the GUI
we can undo these commands easily and read the log to spot this.

From the console we do not care and probably miss that info anyway
as the screen scrolls by.