man: document the use of `!' in the importer
Document the shell invoke in the importer for 18.7 and the people
who will want the installer to run early.  This is really not what
we want to do ever since we moved to a `installer' user and try to
enforce a live environment to help people test the look and feel,
either by first impression for new users or for previewing the new
version prior to installation.