firmware: many change, such wow
A backport half-requested by @Adschellevis.  It doesn't make
a lot of sense to split our current work into bits and pieces
because if we don't push all of these changes now we end up
needing to push them before the EoL date anyway and this is
better to revert and improve.  Most changes have been on the
development track for a while to mature, others are simple
changes.

Needless to say, requires more testing towards 18.1.9 next
week.

* Parse and return to be removed packages for update summary
* Release type change properly updates the repository, summary
* The firmware options can be set via XML files
* Return repository errors in greater detail (4 new error types)
* Add and use API for major upgrades
* Make returned backend JSON a bit more human-readable
* Fix leak of base/kernel update info on package manager updates
* Refactor package manager update summary parsing for speed