ui: funnel hash through firmware-product file #2367
Part one makes it possible to inject branding info from the Makefile
which is not fully complete yet but can always be finished.  The new
hash value can thus be used as a unqiue identifier for page resources
that may be subject to caching.  By using the git hash it allows us
to have this effect on test commits as well as earch working version
as we don't want to track the changes for each file but still get a
good amount of caching.