interfaces: fix false advertising
It's nice that WAN decided to get a global IPv6 regardless
of what dhcp6prefixonly says, but most of the system should
carry on regardless and ignore that address as much as it
can.  It makes troubleshooting easier and less surprising.

(cherry picked from commit 5007150616c98302d52db40574f3c887d19c414e)