pkg: move rc to /etc/rc.opnsense; closes #2416
Move /usr/local/etc/rc to /etc/rc.opnsense in order to be able
to boot without /usr mounted.  Move the /boot files along with
it since they require the same type of /usr/local-escapting.

1. This needs further testing.  :)
2. Inspect rc script for commands depend on /usr tools

Our rc.shutdown stays where it is, if we can shutdown we should be
able to access it.  If not, it's not grabbed via /etc/rc.shutdown.