sidebar - submenu may be cropped due to small height (fix); closes #2865
@fichtner ready for testing (Submenus now open upwards if the space down is not sufficient.)
and by the way....
i need a pro (like you or fabian) or someone else who can optimize the jquery code because it´s memory usage.the calculations of the current postition with offset and scrollTop brings jquery to the performance limit. and unfortunately you can not calculate that, unfortunately.

Perhaps @fabianfrz can help?

i did my best and I like to learn!