unbound: only use IPv6 when enabled and IPv4 is not preferred; closes #2913
A general disable seems reasonable but the IPv4 prefer does not really
mean we should diable IPv6 here.  In the ticket it was indicated that
Unbound does funky stuff in this scenario so to avoid further violation
of POLA it is disabled as well then.

(cherry picked from commit c3b07c466ac2855279f6351e6daf6018db00365c)