services: dhcpv6 override PD range fix; closes #3047
When using dhcpd6 overide, if the user does not enter a 'to' range
then at present the code writes an invalid PD range to dhcpd6.conf.
This change checks to see whether the user has entered a PD range,
if not then no PD range is written to the conf file.