interfaces: DHCP Override MTU Option
It seems DHCP in 11.2 is honouring the ISPs MTU if it is sent. It
also seems there are some ISPs who send a stupid value. This fix
allows the user to ignore the ISP-supplied MTU (or not) with the
default set to ignore for compatibility with the previous behaviour.

PR: https://github.com/opnsense/core/issues/3173
(cherry picked from commit 28796e83a61c9513e7ba8ec380ef14d160de5f83)