probe media before applying new settings, exclude vlan's (align to configure_interface_hardware()). for https://github.com/opnsense/core/issues/3198
We could debate that media is hardware setting and thus should move to configure_interface_hardware(), but seems to be the case for more settings in interface_configure().
Maybe it's an idea to move all of these things to a single point later, so legacy_interface_details() only needs to be called once.

(cherry picked from commit 7aab4a99d3581abbe909c31ffd8cefdc8deab9d9)