filter, aliases. add alias cleanup. previously a delete of an alias would result in the old one still being there. closes https://github.com/opnsense/core/issues/3431
Since all aliases are written to /usr/local/etc/filter_tables.conf, it should be safe to drop all for which a file in aliastables exists but no entry is available in filter_tables.conf anymore.
For debug purposes we log the entries which we delete, although this probably wouldn't be absolutely necessary, it might help to debug potential issues.