Sort interface groups in GUI to match firewall rule order (#3537)
This now makes it easy and predictable to add interfaces to multiple
groups.

Before this change, the interfaces_groups GUI was sorted by the
order that interface groups were added. However, this was not the order
that the actual pf rules would then be generated making it unpredictable
when adding interfaces to multiple groups.

The filter_rules_sort function already took care of the actual pf rule
order.

I also took care of only marking the filter subsystem dirty when needed.
I tested this patch quite a bit on 19.1.8. When updating, users only
need to make one change to the interface groups to have them sorted in
the GUI.

I hope I did not miss anything because this patch is against master. In
it’s current state, the patch can be directly applied to 19.1.x or
master.