configctl: support using as event handler.
Syslog-NG has a convenient program output, which can be used to execute commands on receiving specific data on syslog. This commit adds functionality to configctl which should allow it to be used as a handler for this feature.

Something like the following should execute carp service events, with a latency of 0.5 seconds to prevent flooding the system.

destination ... {
  program("/usr/local/sbin/configctl -e -t 0.5 interface update carp service_status");
}

(cherry picked from commit 3481c16af5d76966f1ee56374e979b8b3b16f8ce)