(Auth) fix expiry and validity for vouchers (#3931)
Always check if voucher is not expired (either
because of never expires or because of expiry
date is in the future) and ensure session timeout
will be the lowest of validity based on the first
usage, the starttime or expiry (if not never expires).
If one of those conditions is not true, reject
authentication.

Fix #3930

(cherry picked from commit 7d72dbdc2858901661fe662ebd4b799940efb87a)
(cherry picked from commit 862a60b7743eb9b51a7ad014702ce76ac28dcba4)