MVC: support inheritance of the ArrayField type, institutionalise the type by adding a method isArrayType() to the basefield, which returns true if the type itself is a ArrayField or one of its descendants.
Conceptually there are two types containers in our model, the standard one (ContainerField), which only acts as a placeholder without logic and the ArrayField type, which understands repetitive structures and comes with its own uuid reference per item and methods supporting addition and removal of entries.

With the application specific field types it can be practical if you could extend this container type as well, so you can add additional methods on a more logical spot in the code tree.

ref https://github.com/opnsense/plugins/issues/1720