tunables: default net.inet.icmp.reply_from_interface to 1.
The helpt text (man icmp) suggests this is a better default in our usecase:
-----------
(boolean) Use the IP address of the interface the
packet came through in for responses to packets which
are not directly addressed to us.  If enabled, this
rule is processed before all others.  By default,
continue with normal source selection.  Enabling this
option is particularly useful on routers because it
makes external traceroutes show the actual path a
packet has taken instead of the possibly different
return path.
-----------

ref https://github.com/opnsense/core/issues/4094