core / mvc: add new config changed event using syshook structure
o in order for this to work properly we need to change when a config backup is made, previously we performed a backup before the fact, now we backup afterwards. which means the top level always represents the current change (and can thus be signaled to an event handler). After upgrade one might lose a single backup file due to this change, but that should be a small price to pay for progress.
o config backup count was defined incorrect (60 instead of 100 according to the gui)
o the syslog-ng event structure is using the existing configd handler and filters relevant events within a small time frame (which prevents flooding configd)

Since the event is loosely coupled, the risk for releasing this into an existing environment should be rather low.

For https://github.com/opnsense/core/issues/4388

sponsored by : Modirum (https://www.modirum.com/)