firmware: upgrade sets into updatelist #4500
While here the API now tells the page what it should do.
We always consume updates first and then tell the user
it is all well or that an upgrade is available.

Errors are shown as well although when the API has a
fatal issue we don't want to try to force a reaction
and instead just log to console.

The major upgrade needs another fold to be made even
quicker than before but for today that is enough.