firmware: when config.xml looks like a factory reset register plugins/release #4500
This way the build can do all sorts of funny things and we will end up
with a consistent config.xml after boot.  For people restoring other
config.xml that is not the case but in this scenario the user is likely
aware of what he or she is doing.