dhcp: on the GUI pages avoid the use of dhcpd_dhcp_configure()
Router advertisement pages were already doing it but the
DHCPv4/6 pages not. That is problematic for the IPv6 case
because there router advertisements would be reconfigured as
well.

This may not be the final form, but we need to see if this
solves the main concern about intermittent connectivity bumps
on GUI operations.

PR: https://github.com/opnsense/core/issues/4642