Firewall - allow manual reply-to configuration
Refactor firewall edit page to allow selection of a reply-to gateway in stead of the single "disablereplyto" option. Since underscores aren't valid for gateway names,
we should be able to use __disable__ here to mark the previous "disablereplyto" setting which we can unravel when saving settings again.

Our page saves either "disablereplyto" or "reply-to" depending on selection.