firewall: update filterlog reader slightly
CARP in IPv4 and IPv6 uses the IP header's "ttl" or "hoplimit" so
that we use the same name for reading it to avoid duplication.
The values are the same in any case.

Change "flowlabel" to "flow" to avoid confusion with "label".

Change IP "version" to "ipversion" and consolidate CARP "version(2)"
into "version".