interfaces: on "dhcp6prefixonly" include tracking interfaces #5086
This way we can get a GUA on a WAN that works anyway due to
IPv6 magic.  Also protect the return of addresses with the
actual existence of the interface, because otherwise the
VIP readings are inaccurate.  interfaces_addresses() still
works in both modes, but worst case won't map aliases.