xmlrpc: support authentication using API keys
authenticate_user() is not being called by anyone else so
move it to xmlrpc.php and extend it to allow for API keys
to work.

Since the authentication in this page is always bound to
local credentials it makes sense to extend the user-based
access with the API key-based access as they don't get in
the way of each other.

PR: https://github.com/travisghansen/hass-opnsense/discussions/6