system: add "latest.log" symlink and make sure new log is created
We shuffle the cron jobs around a bit to let the script start at
minute 1 of every hour which is close enough.

We might consider a soft-update of the link instead of removing
it every hour depending on how tail -f and such work, but for now
use what PHP has to offer.

PR: https://github.com/opnsense/core/issues/4993