system: add opnsense-log helper POC
Wanted to do this every now and then since people were not
familiar with clog but never got around to do it.  Now that
clog is gone it's easier to do so let's write a few lines
and extend as necessary.  :)

It looks like `tail -f' doesn't watch the symlink and instead
tails the target file which makes nightly rotation invisible
and the log starts stalling so it means for any readers of
latest.log it shouldn't be a persistent read.

PR: https://github.com/opnsense/core/issues/4993