interfaces: return $ifname and decouple the search loop
This looks like a typo in the original code as we expect to return
a "wan|lan|optx" thingy.

"Parent" lookup isn't really very useful except for PPP types.
The question is if there is an edge case that would still allow
this to resolve when the other code preceeding it can not.