interfaces: remove previously broken code
Since the code was only fixed in d7bf08da22 and the lookup is
questionable (especially on IPv4 real interface which is vanilla
as opposed to PPP IPv6 shifting).

The problem with e.g. a wan: pppoe0 -> em1 situation is that
if you assign em1 the answer to the query shifts from "wan" to
"opt1" so we would rather miss the situation to resolve "em1"
since the correct interface is "pppoe0" anyway.

Also looking at callers of convert_real_interface_to_friendly_interface_name()
there isn't a PPP-related call in there anyway that would require
this.