Interfaces / LTE - remove more or less dead cruft the 3gstats collects in some rare cases.
Although there probably are a couple of Huawei modems which do report status info as advertised in the script, there are also a lot who don't and maybe deadlocked when pulling for data on a "random" port.
For now we should remove this, if at some point in the future a sensible method would be supported to poll status in a more "vendor independent" manor, we can always revise.