system: cache obfuscated version info from syslog-ng binary #5371
It changes with each build since it has a rebuild date but most
of the time it is not being reinstalled.  Besides this helps testing
and is relatively portable since we don't rely on the actual content.

(cherry picked from commit aa2625998d1ddc9af21da6817b32fbec19c8e10c)