interfaces: kill the gifif-empty side effect #5540
This can only happen from the GUI when we try to inject a new
interface and we don't know the device name yet.  Split this
into a device creation which reserves the name for us and move
that to configd in the next set.  This also has the benefit that
the GUI page now does configuration-after-write which it is
supposed to do.