interfaces: keep boot order but invoke pre and post device hooks always #5540
OpenVPN was put on the "delayed" ("virtual") list in 07b5964599944
but more out of convenience than common sense.  To make sense of it
now check if the assigned device is provided by the plugin system
except for a bridge which is still having a special role in the code.

The role of bridges is somewhat disputed.  We may be able to set them
up earlier detaching them from the older ordering logic but more tests
need to be done in order to pull it off.