interfaces: nothing requires all vips to be regenerated #5540
The side effect from this may be that VIPs disappear from disabled
interfaces which is how it should have been.

Furthermore, remove the configuration of vips from rc.newwanip which
was added for unknown reasons in pfSense a long time ago and got also
copied to rc.newwanipv6 for us during cleanups but never in pfSense.
Especially since we removed find_interface_*() function family we never
risk reading an alias as a primary address and for all intents and
purposes an alias should never appear from an interface except it was
stripped in error by faulty code.  Win-win really.  ;)