Firmware: make the connectivity audit more robust
Change the ping to size 1500 to emulate a typical download
scenario.  Sometimes small pings will respond but not the
larger fetch.

Secondly, do a separate IPv4 and IPv6 update of the repository
depending on whether we got a corresponding address from the
host.  It should clear up the question if IPv4 or IPv6 or both
is broken/defunct/disabled.