interfaces: clean up this portion
Cite the manual page on the MTU requirement of (additional) interfaces.
This isn't the first time we add to a bridge since we make sure that
both interfaces exist which can only happen at the end of the boot sequence.

Speed up the reading of ifconfig values by only obtaining it once.