dhcp: reload action for cron; closes #5410
Change the structure a little to support reloading either all
dhcp-related services or only a certain address family through
optional parameter: "inet" for IPv4 and "inet6" for IPv6.

The issue with the original proposal was that although dhcpd6
was restarted we also need to restart radvd since it is affected
by the same bug #4691 and still seems to be happening in the wild
in FreeBSD 13.

While here expand to full path of pluginctl in an unrelated file.