interfaces: not sure why interface_ppps_configure()... #5637
... would be called a number of times during bootup.  Since it's
a WAN interface that is attached there seems little risk of it
having to be reconfigured multiple times during boot unless the
link is flapping but rc.linkup doesn't react to boot as well so
again not sure what the problem might have been.

Code was commented out in pfSense in 2014 shortly after forking
and eventually removed in 2019.