interfaces: include VIPS for primary IPv4 detection #5742
This was surfaced in 22.1.1 when CARP was treated as an alias as well
leaving setups with "none" IPv4 configuration stranded when calling
get_interface_ip() if the service does not support any explicit listening
on a CARP address (which the function actually does support).  Caveats
apply as to which VIP will win as with the code before.  Best used with
just one CARP address.  IP alias doesn't as much sense since that would
be a static IPv4 configuration.

The latter is the reason why we cannot make a similar change to IPv6
where the primary concept is not equal to "first address in ifconfig"
for (sad) reasons.