MVC - model: leave a note in BaseModel about performance. There is still something to gain, but likely not without adding more complexity.
Currently we will call parseOptionData() for each field method in the list, which shouldn't be needed as the underlying model can't change while parsing. Ideally we would extend SimpleXMLElement to offer a sticky/cached version of the parser or cache results for model targets, the latter has the disadvantage that parseXml() should keep track of the path in the model in order to know what it already did.

As this is one of the last "hotspots" according to the profiler. it feels like a good idea to at least document the opportunity for future reference.

(cherry picked from commit b9a25d9685cf37c7b2005c26899b24a5aa0db069)