interfaces: for specific device edit pages avoid reloading device again #5637
The device reconfigure is done unconditionally before configuring the attached
interface.  If we use reload we will configure the interface again, but if the
interface is disabled we wouldn't trigger the device creation.  So go back to
the former code which didn't exhibit any real issue before we used $reload.