system: exclusive lock on /var/run/booting #5841
As soon as rc.bootup execution ended it is safe to acquire
a shared lock on the file making it obsolete to check for
its existence with its previous caveats.

For the time before rc.bootup just assume that nothing
can happen around dynamic network backend reloads as
interfaces are only configured inside rc.bootup.