interfaces: add missing scope to GIF host route; closes #5884
Although we could agree that we should not emit link locals
without scope attached and/or that system_host_route() should
fix this by itself, there is only caller of system_host_route()
with the same problem and that one handles it correctly already.

Since we do not know the side effects and it is late into the
22.7 development with RC1 already out that we fix this issue here
fully with the least impact possible.

(cherry picked from commit 575c8a5233d5867ca708ca9d706217186084a9f4)