system: ensure that _defaultgw files are written #5900
Ok so this might have been an issue in the past WRT DHCP client
and default route hanling which is now done another way but maybe
it shouldn't have.  For now just see how this works in practice
and then decide later if more needs to be done or not.