system: deal with ifctl internals for now #5933
Later we want ifctl to add the scope to the link-local gateway
by default which will remove a more code, but for now we are
bound to undo the internal magic.

Interface name stops with first ":" or last "_".  For the exclude
check use the correct array index, not the array itself.