interfaces: convert lagg and simplify assignment page
list_devices() shouldn't change in the way it's generated
pre- or post-write so just us a pre-write invoke to get the
list of assignable devices.

Not completely true for weird WLAN case with the parent
assignment but sooner or later we will be able to get rid
of that as well.