Core/Firmware - display license validity when applicable
Extends check.sh to fetch license json file when a uuid is found and flush this to /usr/local/opnsense/license.json.
The firmware page adds a license validity row, which is hidden by default and only shown when product.php returns 'product_license_valid_to'.

For the dashboard widget we'll add the "licensed until" string to the core product string when available.