firmware: restructure license information handling
opnsense-update -K will handle subscription key return if found.
Fetch the license metadata if it exists, move it to core version
meta file and handle the contents dymanically if it exists.

The only issue for later is that changing the mirror away from
business we have a lingering license file until the next check
for updates is done.