Services: Captive Portal: Administration - retire mod_evasion (https://github.com/opnsense/core/issues/6332)
As lighthttpd's changelog (https://www.lighttpd.net/2023/1/3/1.4.68/) notes the module is deprecated and can be replaced by mod_magnet with lua script.
Since the firewall offers rate limitting as core feature, we might as well remove the fixed (hard) limit in CP and point people to the firewall rules if needed.