system: yup yup #6376
Most likely this is backend material which needs to be filtered/
used to render page content and react on via formalized API request
in order to avoid leaking data.

The nice thing about the filtering is that we only ever have to look
up service status for matching services making the lookup faster.