configd: cleanup and style fixes, closes https://github.com/opnsense/core/issues/4150
This commit offers support for deeper nested commands by simplifying the call flow of configd. Eventually we should try to make the action types more modular, but before doing that it makes sense to cleanup up what we have now first. Remove csconfigparser.py as we should be able to pass "optionxform" without wrapping the class as well.