interfaces: add "autoconf" to details output
While adding noticed that it wasn't added which was due to
ending interation one cycle too early (last word on the line)
so changed the parsing to slightly.  Reaching forward with
certain keywords in place is safe, just need to make sure to
skip what we have read then.  Same for IPv4 although less
relevant.