firmware: if the kernel hasn't been applied do it now
"opnsense-update -u" changes to not apply the kernel and the upgrade
script will handle it.  But in case this is going to be scripted by
the user we should just make sure this still needs to be taken care of
adding an extra reboot but that's ok.

(cherry picked from commit ea8338767dc73c93460694d0b3d1ece65addd87c)