dhcp: tighten expire logic for delegated prefix routes
We need to expire existing routes as well in order to get rid
of the old gateway, because deleting with the new gateway given
will not delete the existing route to another one.

Also bring back (and actually use) the active binding check
before storing an address or prefix for processing.

Use "null" to designate that these are expired, later overwritten
if an active lease is there.