MVC/Model/ArrayField - add generic static record definition.
While converting legacy code we do seem to have multiple static entries in different models, in which case we tend to overwrite ArrayField and extend the output with the static entries that are required for the model. These can either be defined in the legacy code (SPDField) or defined as static options (like the bogons in AliasField).

This commit adds a method getStaticChildren() which the derived class may overwrite in order to spawn new items to the list which can not be persisted. This should help avoid duplication of similar code constructs in various migrated areas. These items are constructed during the "post loading" event.