interfaces: introduce a lock and DAD timer into newwanip for IPv6 #6637
There seem to be two issues:

1. Tentative addresses could have always been ignored for the wrong reasons
and we can savely move the delay to this script even though a small delay
will be the result (2 seconds with the default sysctl).  Not sure why this
problem previously not mattered that much, but at least we can move the other
instance of the delay to here and avoid duplication since it will continue
to load this script anyway.

2. Due to overlaps and technical convolution these scripts can be run multile
times in a very short succession especially on a bootup.  Since we have a delay
here now we force a lock prior to "catch" stray invocations.  The only issue
I see is that we could lose the "force" flag in the process, but if that is
the case the log message will reveal and we can work around this sas well with
a two stage log perhaps.

(cherry picked from commit 2be79db6f8dcdcfd1e23d7c197bfcec61c8dcd61)