wizard: restrict to validating IPv4; closes #6681
The system wizard is IPv4 only but doesn't use the correct
function.  It's debatable if we need to restrict DNS servers
but if all of this is only IPv4 we should stick to it in a
consistsent way.

(cherry picked from commit c9bd89752bc76c9d4112e12cdd7c308330df835f)