interfaces: adjust MTU before creating IPv4 config
This is possible now since we no longer set the MTU on the PPP device,
wich is not created before IPv4 handling.  We may have to add the MTU
to the PPPoE device as well, but from the looks of it this does not
matter for one reason or another.