interfaces: avoid duplicating $vfaces array
The idea of list_interfaces.php was good so let's take it further
and inline the use into legacy_interfaces_details() and moving the
patterns to a shared function.  Also add a device name hint in case
the array key is not passed for some reason which can help when
inspecting debug data.  Since legacy_interfaces_details() is already
available in pluginctl -D replace the action command and drop the
obsolete script.