system: kill the unused "alert" "interval"
The terminology is confusing -- it's just a poll interval in
dpinger's alert thread like we have in gateway watcher (also
defaulting to 1 second).  Since we do not want to expose
such values to the GUI without a pressing reason we better
do it this way than add more knobs that make support harder
and expectations less clear (reaction time suffers quickly when
increasing this value).

While here disable the alert thread completely in dpinger which
seems to be steered by passing the loss and delay.  Also adjust
the default for the loss interval which is calculated by dpinger
as 4 times the probe interval, so the default was wrong also.

(cherry picked from commit 97b08eec0c594a2ff953c32fd5cac55f93fd2be7)