unbound: allow disabled interfaces in interface field
While we are not concerned with interfaces that have disappeared
we should keep track of disabled interfaces since there is no strict
link between this field and the interface (or OpenVPN) configuration
which would make validation fail for the wrong reasons and/or force
the user to deselect the interface in order to proceed.

Also exclude lo0 because it is always added to ensure local resolving
which had no complaints since 2018 when this was reworked.  The abilty
to select it would cause the question of why it's not adhered to if
anyone started looking.