mvc: throw on template reload issues in mutable service controller #6487
While this omits all other callers of "template reload" functionality
we would only like to be concerned with these GUI ones and migrate everything
to using a mutable controller (for maintenance reasons at the very last).

Only make a short hint to where the error is.  The service is not started
on purpose now.