mvc: emit correct message on required validation in BaseField
The base field would emit the validation message of the derived
field, but that is strange sometimes since the actual reason is
much more simple and less ambiguous.  Also makes it able to trace
the error to the model definition of Required=Y instead of routing
it through a different field that may have had no say in this.

Noticed on pools name input that throws "text validation error",
but was really complaining about the lack of a value.

(cherry picked from commit bcd8cec574196be8ca6c24df941c192a588c37a4)