mvc: remove special validation messages likely never seen
We could use defaultValidationMessage() but the BaseListField message
is fitting enough and the GUI itself cannot cause this to happen.
The reason for stripping is also that the variable is not handled by
the field so there is no special validation meaning attached either.

For BaseField derivates the idea was to set a custom required validation
message which is no longer possible or possibly not even reachable.